// layers/gps/server/schemas/IBodyMqtt.ts
export interface IBodyMqtt {
  state: {
    reported: {
      16: number
      21: number
      66: number
      67: number
      68: number
      69: number
      181: number
      182: number
      200: number
      239: number
      240: number
      241: number
      ts: number
      pr: number
      lat: number
      lng: number
      latlng: string
      alt: number
      ang: number
      sat: number
      sp: number
      evt: number
    }
  }
}